refaktorere rydder opp funksjoner
