+++ apps/etterlatte-behandling/src/main/kotlin/behandling/jobs/EtteroppgjoerJobService.kt
@@ -4,10 +4,12 @@ import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.newSingleThreadContext
 import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.behandling.etteroppgjoer.EtteroppgjoerService
+import no.nav.etterlatte.behandling.etteroppgjoer.EtteroppgjoerStatus
 import no.nav.etterlatte.behandling.etteroppgjoer.EtteroppgjoerToggles
 import no.nav.etterlatte.behandling.etteroppgjoer.forbehandling.EtteroppgjoerForbehandlingService
-import no.nav.etterlatte.behandling.etteroppgjoer.sigrun.SkatteoppgjoerHendelserService
+import no.nav.etterlatte.behandling.klienter.VedtakKlient
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggleService
+import no.nav.etterlatte.libs.ktor.token.HardkodaSystembruker
 import org.slf4j.LoggerFactory
 import java.time.YearMonth
 
@@ -15,7 +17,7 @@ import java.time.YearMonth
 class EtteroppgjoerJobService(
     private val etteroppgjoerService: EtteroppgjoerService,
     private val etteroppgjoerForbehandlingService: EtteroppgjoerForbehandlingService,
-    private val skatteoppgjoerHendelserService: SkatteoppgjoerHendelserService,
+    private val vedtakKlient: VedtakKlient,
     private val featureToggleService: FeatureToggleService,
 ) {
     private val logger = LoggerFactory.getLogger(this::class.java)
@@ -26,31 +28,65 @@ class EtteroppgjoerJobService(
             runBlocking(ctx) {
                 if (featureToggleService.isEnabled(EtteroppgjoerToggles.ETTEROPPGJOER_PERIODISK_JOBB, false)) {
                     logger.info("Starter periodiske jobber for etteroppgjoer")
-                    startEtteroppgjoerKjoering(etteroppgjoerService, etteroppgjoerForbehandlingService, skatteoppgjoerHendelserService)
+                    startEtteroppgjoerKjoering()
                 } else {
                     logger.info("Periodisk jobber for etteroppgjoer er deaktivert")
                 }
             }
         }
     }
-}
 
-internal fun startEtteroppgjoerKjoering(
-    etteroppgjoerService: EtteroppgjoerService,
-    etteroppgjoerForbehandlingService: EtteroppgjoerForbehandlingService,
-    skatteoppgjoerHendelserService: SkatteoppgjoerHendelserService,
-) {
-    val yearNow = YearMonth.now().year
-    val aarMellom2024OgNaa = (2024..yearNow).toList()
+    fun startEtteroppgjoerKjoering() {
+        val yearNow = YearMonth.now().year
+        val aarMellom2024OgNaa = (2024..yearNow).toList()
 
-    val antallHendelser = 500
+        for (inntektsaar in aarMellom2024OgNaa) {
+            finnOgOpprettEtteroppgjoer(inntektsaar)
 
-    for (inntektsaar in aarMellom2024OgNaa) {
-        etteroppgjoerService.finnOgOpprettEtteroppgjoer(inntektsaar)
+            // TODO: legge inn denne når vi har testet litt mer?
+            // skatteoppgjoerHendelserService.startHendelsesKjoering(HendelseKjoeringRequest(500),"automatisk")
 
-        // TODO: legge inn denne når vi har testet litt mer?
-        // skatteoppgjoerHendelserService.startHendelsesKjoering(HendelseKjoeringRequest(500),"automatisk")
+            finnOgOpprettForbehandlinger(inntektsaar)
+        }
+    }
 
-        etteroppgjoerForbehandlingService.finnOgOpprettForbehandlinger(inntektsaar)
+    // finn saker med etteroppgjoer og mottatt skatteoppgjoer som skal ha forbehandling
+    fun finnOgOpprettForbehandlinger(inntektsaar: Int) {
+        logger.info(
+            "Starter oppretting av forbehandling for etteroppgjør med mottatt skatteoppgjør for inntektsår $inntektsaar",
+        )
+        val etteroppgjoerListe =
+            etteroppgjoerService.hentEtteroppgjoerForStatus(EtteroppgjoerStatus.MOTTATT_SKATTEOPPGJOER, inntektsaar)
+
+        for (etteroppgjoer in etteroppgjoerListe) {
+            try {
+                etteroppgjoerForbehandlingService.opprettEtteroppgjoerForbehandling(
+                    etteroppgjoer.sakId,
+                    etteroppgjoer.inntektsaar,
+                    HardkodaSystembruker.etteroppgjoer,
+                )
+            } catch (e: Exception) {
+                logger.error("Kunne ikke opprette forbehandling for sakId=${etteroppgjoer.sakId} grunnen: ${e.message}")
+            }
+        }
+    }
+
+    // finn saker som skal ha etteroppgjør for inntektsår og opprett etteroppgjør
+    fun finnOgOpprettEtteroppgjoer(inntektsaar: Int) {
+        no.nav.etterlatte.logger
+            .info("Starter oppretting av etteroppgjør for inntektsår $inntektsaar")
+        val sakerMedUtbetaling =
+            runBlocking {
+                vedtakKlient.hentSakerMedUtbetalingForInntektsaar(
+                    inntektsaar,
+                    HardkodaSystembruker.etteroppgjoer,
+                )
+            }
+
+        sakerMedUtbetaling
+            .forEach { sakId -> etteroppgjoerService.opprettEtteroppgjoer(sakId, inntektsaar) }
+
+        no.nav.etterlatte.logger
+            .info("Opprettet totalt ${sakerMedUtbetaling.size} etteroppgjoer for inntektsaar=$inntektsaar")
     }
 }
